BUG/MINOR: quic: Possible stream truncations under heavy loss
This may happen during retransmission of frames which can be splitted
(CRYPTO, or STREAM frames). One may have to split a frame to be
retransmitted due to the QUIC protocol properties (packet size limitation
and packet field encoding sizes). The remaining part of a frame which
cannot be retransmitted must be detached from the original frame it is
copied from. If not, when the really sent part will be acknowledged
the remaining part will be acknowledged too but not sent!

Must be backported to 2.7 and 2.6.

(cherry picked from commit dd419461eff5395b369f5497fef703dc383577bb)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 42331e27db41499fd2f4a08bacca6bbe468debcd)
Signed-off-by: Willy Tarreau <w@1wt.eu>