BUG/MEDIUM: sink: Fix release of sinks during the deinit
A regression was introduced when the commit db5fd2b013 ("BUG/MINOR: sink:
free the forwarding task on exit") was backported. The call to ring_free()
in sink_deinit() was erroneously removed. The result is a memory corruption
when HAProxy exists that can lead to a crash.

It is a 2.2-specific commit. Thus there is no upstream ID. No backport
needed.