MINOR: quic: When probing Handshake packet number space, also probe the Initial one
This is not really a bug fix but an improvement. When the Handshake packet number
space has been detected as needed to be probed, we should also try to probe the
Initial packet number space if there are still packets in flight. Furthermore
we should also try to send up to two datagrams.

Must be backported to 2.6 and 2.7.

(cherry picked from commit 37ed4a3842d0f8c4a4f77c5f57b2dc1f0401a7e1)
Signed-off-by: Willy Tarreau <w@1wt.eu>