MINOR: quic: display CIDs and state in "show quic"
Complete "show quic" handler. Source and destination CIDs are printed
for every connection. This is complete by a state info to reflect if
handshake is completed and if a CONNECTION_CLOSE has been emitted or
received and the allocation status of the attached MUX. Finally the idle
timer expiration is also printed.

This should be backported up to 2.7.

(cherry picked from commit 58d9d5d1607635c8f47b1c398ae6ca70312fda10)
Signed-off-by: Willy Tarreau <w@1wt.eu>