BUG/MINOR: mux-h1: Don't report an H1C error on client timeout
When a client timeout is reported by the H1 mux, it is not an error but an
abort. Thus, H1C_F_ERROR flag must not be set. It is espacially important to
not inhibit the send. Because of this bug, a 408-Request-time-out is
reported in logs but the error message is not sent to the client.

This patch must be backported to 2.7.

(cherry picked from commit 3eff752f6f157985e92cd9d9ca3da27ee0af2772)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>