BUG/MINOR: proto_ux: report correct error when bind_listener fails
In uxst_bind_listener() and uxdg_bind_listener(), when the function
fails because the listener is not bound, both function are setting
the error message but don't set the err status before returning.

Because of this, such error is not properly handled by the upper functions.

Making sure this error is properly catched by returning a composition of
ERR_FATAL and ERR_ALERT.

This could be backported up to 2.4.

(cherry picked from commit d861dc9b48789f9c6c7f092d1a3114695cfbbe0d)
Signed-off-by: Willy Tarreau <w@1wt.eu>